test(e2e): one project-creation helper for the three-step wizard
WHY THIS MATTERS
The New Project wizard became Details -> Access -> Add Documents with a
type="button" "Create project" primary, but only one of the three specs
that create projects was updated. critical-path.spec.ts clicked one
Next and waited for an Upload button that exists only on the documents
step, then clicked button[type="submit"]; chat-management.spec.ts
waited for /create project|creating/i on the Access step. Both would
fail on every CI run. Separately, the new aria-label "Actions for
<title>" on the sidebar menu trigger made two getByRole name lookups
match two buttons (substring matching), a strict-mode violation.

HOW THE FIX WORKS
e2e/projects.ts exports createProject(page, name, filePath?) and all
three specs use it, so a fourth wizard step cannot be missed again. The
two ambiguous locators pass exact: true; the aria-label stays because
it is a real accessibility improvement.
